Refactor time selectors state shape to use array
Why:
Previously we were using a complex and nested object to encode selector
state in the application. This is both more structure than what is
required, increasing mental overhead to understand, and mixes the
concerns of updatable state and static presentational data, introducing
unnecessary coupling between components and the redux store. (We don't
need redux or a web server to tell us there is 7 days in the week and
what their names are.)

This commit:
Refactors the time selector components and their slice of redux state to
use an array to encode selected days. Components are provided constants
to supply button title information. These titles will also be used as
the toggle selector action payloads.

In general:
- Redux should only contain information which can change during app use.
- Components should stand on their own and not depend on redux to
render a default state.
- The redux should be as flat as possible, nested data is a code smell.