Refactor styling in components
Why:
We would like to have consistent and lean styling in the app. In general
we would like to use patterns that are ergonomic (easy to use,
understand, and update) and expressive (declarative rather than
imperative or prescriptive)

This commit:
Refactors a lot of styling implementation by removing unused code,
reorganizing some code, and pulling common button logic into a buttons
style module.