Replace redux with react context api
Why:
Currently, we are using redux to handle state shared between components.
Given the simplicity of the application, redux is excessive and the
functionality it provides could be handled by using the react context
api, with less machinery. Additionally, we would like to remove as many
dependencies as possible from the template app, to both reduce the
maintenance burden and make the app more generally applicable to all
react native projects.

This commit:
Removes all use of redux components and patterns within the app,
replacing the functionality with a react context provider and consumers.