Update to rn 63, rename, and add typescript
Why:
There are three significant updates we would like to make to the base
setup of this project:

1. Use the most recent version of react-native

As a matter of best practice, we would like to keep the repo up to date
with the ecosystem's core libraries. We are currently on
react-native 0.61 and would like to upgrade to 0.63

2. Change the name of the project to ReactNativeTypescriptStylesExample

Currently, the app is named 'RNStyleingTemplate.' This name is
suboptimal for SEO purposes and misrepresents the intended use case of
an example of style application for react-native components. That is we
expect most internet searches to use the term 'React Native,' not 'RN,'
and for users of the project to review the project for reference, not
use the project as a starter template.

3. Introduce TypeScript

thoughtbot recommends TypeScript for all React Native projects.

This commit:
Updates the project to use react-native 63, renames the project from
'RNStyleingTemplate' to 'ReactNativeTypescriptStylesExample' and
introduces typescript to the build chain.

We achieved these three items by running the react-native CLI init tool
with the typescript template and porting that boilerplate code to the
repo, leaving the application logic in src unchanged. While this is a
lot of changes to be done in a single commit, this is the most
time-efficient way to make the desired changes. This commit does not
introduce any changes to the application logic.