Update readme and refactor components
This commit:
Updates the readme to reference the styles repo and to show a few
example images of the example app. We also refactored the screen
component by pulling out the base example into its own file and added
some extra margin such that a single section takes up the full height of
a large format phone.